requirements.txt
beautifulsoup4>=4.12.0
pandas>=2.2.0
PyYAML>=6.0.1
requests>=2.31.0
streamlit>=1.35.0
